decimal functionality added by doubles, learned that array.convertall works not just for int but for double and long, good deal

+++ reportCardFormatter/Program.cs
@@ -4,8 +4,8 @@ public class Program
     //global variables
 
     public static string[] userInput= {"0", "0", "0",};
-    public static int[] userGrades= {0, 0, 0,};
-    public static int userAverage = 0;
+    public static double[] userGrades= {0, 0, 0,};
+    public static double userAverage = 0;
 
     public static string userName = "john-jane";
 
@@ -29,7 +29,7 @@ public class Program
 
         // I had an idea to do this in swoop and the internet said yes use this method 
         // Converts the entire array in one line
-        userGrades = Array.ConvertAll(userInput, int.Parse);
+        userGrades = Array.ConvertAll(userInput, double.Parse);
         //do average
         userAverage = (userGrades[0] + userGrades[1] + userGrades[2])/3;
